Moved CustomTabPixelNames to app.
